@@ -1,4 +1,10 @@
 import React from 'react'
 import { render } from 'react-dom'
-import App from './modules/App'
-render(<App/>, document.getElementById('app'))
+import { Router, Route, browserHistory, IndexRoute } from 'react-router';
+import routes from './modules/routes';
+
+render(
+    (
+    <Router history={browserHistory} routes={routes} />
+    ), document.getElementById('app')
+  )
